+++ src/activation.h
@@ -1,6 +1,6 @@
 #ifndef ACTIVATION_H
 #define ACTIVATION_H
 
-int apply_relu(double *input, int size);
+int apply_relu(float *input, int size);
 
 #endif
